TrafficSign-Recognition-System/requirements.txt
fastapi
uvicorn[standard]
numpy
pandas
Pillow
keras
tensorflow
jinja2
requests
matplotlib